prothomalo_scraper_project/requirements.txt
django
elasticsearch
celery
redis
djangorestframework
requests
beautifulsoup4
